Avoid polluting the the x11 shortcut map.
When user presses a key, but this key was not previously registered as a
shortcut, the shortcut map is polluted with "" entry.

This could cause an error when the shortcut is not set after clicking OK in the
keybinging editing dialog and you are told that shortcut "" is being used
instead (in the logs).

Signed-off-by: Roman Kapl <code@rkapl.cz>